Three-body reactions and Merged PlotRates/PlotRatesGroup
Added capability to manage 3-body reactions (basically another column in Gstr/iG). Thanks Pam Rickly for assistance.

Overhauled family identification in PlotRatesGroup and merged back in with PlotRates. PlotRates will now work for a single species or a family Thanks Sam for pointing out the bug that lead to this update.

Updated documentation for v4.1.